Add BlackBerryLoginActivity to cover the web login
The activity should be probably renamed to WebLoginActivity.
It is not very complete but works. Any connection problems
are not tested and probably not very well handled (I do not
know how to do it properly).

The login data are logged using Log.d which may leak some
sensitive information (should be switched off later).

Principle:

* start WebView and load Google login page
* user inserts its credentials and logs into his/her account
* then the user accepts the application's access (done when errorCode == -2)
* we now obtain the token by FinishLoginTask and provide it to
  the calling activity as data identified by "google.token" string