Implement Controller.assignToken(), alternative to login()
When login is not performed by the standard Google way,
we obtain the token by another channel. This is the case
of BlackBerryLoginActivity. So, we start with token only
and obtain Google User Info from the server.

No persistence of user data is implemented at the moment.
I would be glad to have it but the spagetti code of the
login and doLogin functions is quite unreadable to do
it easily. It is not clear when/how to check the persistable
storage before the actual login... To be done later.