Fix doing network requests from more threads
With this commit is doRequest method synchronized which means it will block other threads until return which means request and response will be done together. This way only one request will be sent to server at once.
Without it it caused getting wrong responses to wrong requests which resulted in crashes...