Fix for not showing progressbar on some occasions
When caller starts the new task in onExecute callback of existing task, then the "old" progressbar was hidden after new task started (which showed "new" progressbar), and thus the old callback hid the "new" progressbar.

Now onExecute callback is started at the end of the method which solves this problem.